+++ R/get_fasta_header.R
@@ -33,7 +33,9 @@ get_fasta_header <- function(
 
   withr::local_envvar(
     .new = list(
-      BLAST_USAGE_REPORT = "false"
+      BLAST_USAGE_REPORT = "false",
+      NCBI_DONT_USE_NCBIRC = "true",
+      NCBI_DONT_USE_LOCAL_CONFIG = "true"
     ),
     action = "replace"
   )